requirements.txt
fastapi
uvicorn
python-multipart
jinja2
openai
openai-whisper
python-dotenv
ffmpeg-python
fuzzywuzzy
python-Levenshtein
python-telegram-bot
langdetect
twilio
aiofiles
requests
